Split RandomMixin into prng and seed classes
