+++ src/interfaces/swig/Distribution.i
@@ -26,6 +26,7 @@
 %include <shogun/distributions/Distribution.h>
 
 /** Instantiate RandomMixin */
+%template(SeedableDistribution) shogun::Seedable<shogun::CDistribution>;
 %template(RandomMixinDistribution) shogun::RandomMixin<shogun::CDistribution, std::mt19937_64>;
 
 %include <shogun/distributions/Histogram.h>
