+++ src/interfaces/swig/RandomMixin.i
@@ -1,5 +1,7 @@
 %{
+ #include <shogun/mathematics/Seedable.h>
  #include <shogun/mathematics/RandomMixin.h>
 %}
 
+%include <shogun/mathematics/Seedable.h>
 %include <shogun/mathematics/RandomMixin.h>
