+++ src/interfaces/swig/Statistics.i
@@ -32,6 +32,7 @@
 %include <shogun/statistical_testing/TwoSampleTest.h>
 
 /** Instantiate RandomMixin */
+%template(SeedableTwoSampleTest) shogun::Seedable<shogun::CTwoSampleTest>;
 %template(RandomMixinTwoSampleTest) shogun::RandomMixin<shogun::CTwoSampleTest, std::mt19937_64>;
 
 %include <shogun/statistical_testing/MMD.h>
